+++ frontend/src/App.vue
@@ -1,0 +1,9 @@
+<script setup>
+import AppLayout from './components/AppLayout.vue';
+import { RouterView } from 'vue-router';
+</script>
+<template>
+  <AppLayout>
+    <RouterView />
+  </AppLayout>
+</template>